+++ test/11-custom_elements.js
@@ -511,7 +511,7 @@ describe('CustomElement', function() {
 
 			let result = await renderWithPledge(compiled);
 
-			assertEqualHtml(result.html, '<template-slot-test he-rendered="1"><div data-he-slot="main">This will set the content of the <b>main</b> slot</div></template-slot-test>');
+			assertEqualHtml(result.html, '<template-slot-test he-rendered="1"><slot name="main"><div slot="main">This will set the content of the <b>main</b> slot</div></slot></template-slot-test>');
 		});
 
 		it('should not confuse slots with similar elements', async function() {
@@ -527,7 +527,7 @@ describe('CustomElement', function() {
 
 			let result = await renderWithPledge(compiled);
 
-			assertEqualHtml(result.html, '<template-slot-test he-rendered="1"><div data-he-slot="main">Slot test 1</div></template-slot-test>\n<template-slot-test he-rendered="1"><div data-he-slot="main">Slot test 2</div></template-slot-test>');
+			assertEqualHtml(result.html, '<template-slot-test he-rendered="1"><slot name="main"><div slot="main">Slot test 1</div></slot></template-slot-test> <template-slot-test he-rendered="1"><slot name="main"><div slot="main">Slot test 2</div></slot></template-slot-test>');
 		});
 
 		it('should render the contents after the attributes have been set', async function() {