+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Add support for ternary conditionals
 * Add `CustomElement.defineStateVariable(name, config)` static method
 * Add `CustomElement#setState(name, value)` and `CustomElement#getState(name)` methods
+* Make *template slots* behave more like vanilla custom element slots
 
 ## 2.3.19 (2024-04-13)